+++ alembic/versions/2023_10_05-add_bounding_box_to_cities_table_c858663fb7c1.py
@@ -9,7 +9,7 @@ from alembic import op
 import sqlalchemy as sa
 from sqlalchemy.orm.session import Session
 
-from brew_scout.libs.dal.models.cities import CountryModel, CityModel
+from brew_scout.libs.dal.models.shops import CountryModel, CityModel
 from brew_scout.libs.domains.cities import City
 
 # revision identifiers, used by Alembic.
