+++ alembic/versions/2024_04_22-add_and_update_bounding_boxes_9ec756593897.py
@@ -9,7 +9,7 @@ from alembic import op
 import sqlalchemy as sa
 from sqlalchemy.orm import Session
 
-from brew_scout.libs.dal.models.cities import CountryModel, CityModel
+from brew_scout.libs.dal.models.shops import CountryModel, CityModel
 from brew_scout.libs.domains.cities import City
 
 # revision identifiers, used by Alembic.
